Backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
decorator==5.1.1
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.63.6
django-auditlog==2.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-ipware==6.0.4
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
et-xmlfile==1.1.0
executing==2.0.1
fpdf==1.7.2
graphviz==0.20.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
ipython==8.23.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.6
msgpack==1.0.8
mssql-django==1.4.2
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
parso==0.8.4
phonenumbers==8.13.40
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyodbc==5.1.0
pyOpenSSL==24.1.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==2.0.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
redis==5.0.4
referencing==0.34.0
reportlab==4.2.0
requests==2.32.2
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==5.0.2
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.0
zope.interface==6.3
